port h096 infrastructure improvements to master
Generic DAQ/analysis improvements from h096_31Si_dp:

DAQ scripts (Pi/DAQ side):
- start_run.sh: AI mode (--ai flag), HV on check, safety reset, elog_done signal
- stop_run.sh: AI mode, use daqDataPath, write lastRunComment/totalFileSize to expName.sh
- AutoStartStop: replace sleep 10 with elog_done signal wait (max 30 s)
- AutoTuneThreshold.py: Discord + RDT support, refactored onto common.py
- DataBase.py: Python3 syntax fixes (print, semicolons)
- WriteComment: rename GeneralHelios -> MasterHelios

New DAQ scripts (Pi):
- common.py: shared Python parser library for SetChannel/GetChannel/AutoTune
- SetChannel: pyepics channel setter for 4-side array (replaces SetVMEThreshold)
- GetChannel: Python3 mirror of SetChannel, reads mapping dynamically
- GenElogExtra.py: detector threshold + HV status table for elog

Removed: SetVMEThreshold (superseded by SetChannel)
Moved to obsolete/: AutoMonitor, HVMonitor, dig_setup*, gaugeDB, globus_*, helios_setup_digitizer, ryan_testing, terminals, timer.sh, trigonly_setuphelios

Mac-side workflow (new):
- start_run_Mac.sh: SCP expName.sh, GenElog, GenElogExtra, post to elog + Discord
- stop_run_Mac.sh: Grafana screenshot, GenElog stop entry, post to elog + Discord
- GenElog.py: generate elog entry from expName.sh
- GenElogExtra.py: detector status table
- common.py: shared Python parser
- push2Discord.sh: HTML -> Discord Markdown conversion, fix empty RunNum
- push2Elog.sh: elogFull.txt, CloudFlare elog bin, fix ID parsing
- GrafanaElog.sh/GrafanaWeb.sh: rename GeneralHelios -> MasterHelios

tcpReceiver.cpp: move GEB_TYPE_DGS/MAX_FILE_SIZE to #define, simplify args, sprintf->snprintf

Analysis:
- HELIOS_LIB.h: angular straggling impl, NaN fix in CalHit (sign=-1 hardcoded), GetDeltaTheta()
- BinaryReader.h: mmap fast path + memory leak fix (munmap/close in destructor)
- EventBuilder_A.cpp: new EventBuilder variant (A-style sorting)
- makefile: add EventBuilder_A target; allow EventBuilder*.cpp in .gitignore
- AnalysisLibrary.h: fix signed/unsigned comparison warnings

expName.sh: add lastRunComment/lastStopComment/totalFileSize fields (written by stop_run.sh)
.gitignore: add daq/saves/trigger.txt, daq/null, daq/trigger, __pycache__, analysis/working_Helios/

+++ analysis/Armory/AnalysisLibrary.h
@@ -129,7 +129,7 @@ std::vector<std::vector<float>> LoadFitParameters(std::string fileName, bool pri
       if( str.size() < 2 ) continue;
 
       int detID = atoi(str[0].c_str());
-      if( parameters.size() <= detID ) parameters.resize(detID+1);
+      if( (int) parameters.size() <= detID ) parameters.resize(detID+1);
       
       for( size_t j = 1 ; j < str.size() ; j++ ) {
         parameters[detID].push_back( atof(str[j].c_str()) );
@@ -142,8 +142,8 @@ std::vector<std::vector<float>> LoadFitParameters(std::string fileName, bool pri
 
   if (print) {
     printf("Loaded fit parameters from file: %s | array size = %zu\n", fileName.c_str(), parameters.size());
-    for( int i = 0 ; i < parameters.size() ; i++ ) {
-      printf("detID %2d: ", i);
+    for( size_t i = 0 ; i < parameters.size() ; i++ ) {
+      printf("detID %2zu: ", i);
       for( size_t j = 0 ; j < parameters[i].size() ; j++ ) {
         printf("% .6e ", parameters[i][j]);
       }
